// tools/firebase/functions-ts/src/enums/transactions/transaction-reason.ts
export enum TransactionReason {
  DEPOSIT = 'DEPOSIT',
  WITHDRAWL = 'WITHDRAWL',
  TASK_PUBLISH_DEDUCTION = 'TASK_PUBLISH_DEDUCTION',
  BONUS_RECEIVED = 'BONUS_RECEIVED',
  OFFERING_REWARD = 'OFFERING_REWARD',
  TASK_DELETED_REFUNED = 'TASK_DELETED_REFUNED',
  TRANSFER_TO_SPACE_WALLET = 'TRANSFER_TO_SPACE_WALLET',
  LEADS_APP_PLAN_PURCHASE = 'LEADS_APP_PLAN_PURCHASE',
  LEADS_APP_TASK_CREDIT_PURCHASE = 'LEADS_APP_TASK_CREDIT_PURCHASE',
  NIKAT_APP_ORDER_PAYMENT = 'NIKAT_APP_ORDER_PAYMENT',
  CUSTOM_WEBSITE_PURCHASE = 'CUSTOM_WEBSITE_PURCHASE',
  APP_PRODUCT_PURCHASE = 'APP_PRODUCT_PURCHASE',
  SUBSCRIPTION_PAYMENT = 'SUBSCRIPTION_PAYMENT'
}